Octodino

+++ easyfile/index.js
@@ -1,3 +1,6 @@
+
+//https://nodejs.org/api/fs.html#filehandlereadoptions
+
 //Include path libary
 var path = require('path');
 //Include fs libary
@@ -5,16 +8,24 @@ var fs = require('fs');
 
 //Create new file
 function create(dirname, fileName, fileContent) {
-fs.writeFile(path.join(dirname, fileName), fileContent, function(err) {
-        if (err) {
-            return console.error(err);
-        }
-        var output = "File " + fileName + " was created.";
-        return output;
-    })
+    if(fileContent == null){
+        console.error("No file content given!");
+        fileContent = "";
+    }
+    fs.writeFile(path.join(dirname, fileName), fileContent, function(err) {
+            if (err) {
+                return console.error(err);
+            }
+            var output = "File " + fileName + " was created.";
+            return output;
+        })
 };
 
 function write(dirname, fileName, fileContent) {
+    if(fileContent == null){
+        console.error("No file content given!");
+        fileContent = "";
+    }
     fs.writeFile(path.join(dirname, fileName), fileContent, function(err) {
         if (err) {
             return console.error(err);
@@ -25,7 +36,15 @@ function write(dirname, fileName, fileContent) {
 };
 
 //Delete file
+/*TODO: Not Production ready!
 function deleteFile(dirname, fileName) {
+    //look if file exists
+    if(exists(dirname, fileName)){
+        console.error("File don't exist in this directory!");
+        var output = "File don't exist in this directory!";
+        return output;
+        //break funcion
+    }
     fs.unlink(path.join(dirname, fileName), function(err) {
         if (err) {
             return console.error(err);
@@ -34,8 +53,10 @@ function deleteFile(dirname, fileName) {
         return output;
     })
 };
+*/
 
 //read file for a specific word
+/* TODO: Not Production ready!
 function readFor(dirname, fileName, searchedContent) {
     fs.readFile(path.join(dirname, fileName), 'utf8', function(err, data) {
         if (err) {
@@ -45,10 +66,11 @@ function readFor(dirname, fileName, searchedContent) {
         return output;
     })
 };
+*/
 
 //look if file exists
 function exists(dirname, fileName) {
-    fs.exists(path.join(dirname, fileName), function(exists) {
+    /*fs.exists(path.join(dirname, fileName), function(exists) {
         if (exists) {
             var output = true;
             return output;
@@ -57,21 +79,43 @@ function exists(dirname, fileName) {
             return output;
         }
     })
+    */
+    try {
+        if (fs.existsSync(dirname + '/' + fileName)) {
+          //file exists
+          var output = true;
+            return output;
+        }else{
+            var output = false;
+            return output;
+        }
+      } catch(err) {
+        console.error(err)
+      }
 }
 
 //copy a file from dirname to targetDirname
-function copy(dirname, fileName, targetDirname) {
-    fs.copyFile(path.join(dirname, fileName), path.join(targetDirname, fileName), function(err) {
+/*TODO: Not Production ready!
+function copy(dirname, fileName, targetDirname, newFileName) {
+    fs.copyFile(path.join(dirname, fileName), path.join(targetDirname, newFileName), function(err) {
         if (err) {
             return console.error(err);
         }
-        var output = "File " + fileName + " was copied.";
+        var output = "File " + fileName + " was copied to " + targetDirname + " with the name " + newFileName + ".";
         return output;
     })
 }
+*/
 
 //rename a file
 function rename(dirname, fileName, newFileName) {
+    //look if file exists
+        if(exists(dirname, newFileName)){
+            console.error("File already exist in this directory!");
+            var output = "File already exist in this directory!";
+            return output;
+            //break funcion
+        }
     fs.rename(path.join(dirname, fileName), path.join(dirname, newFileName), function(err) {
         if (err) {
             return console.error(err);
@@ -82,6 +126,7 @@ function rename(dirname, fileName, newFileName) {
 }
 
 //get the full data of a file
+/* TODO: Not Production ready!
 function read(dirname, fileName) {
     fs.readFile(path.join(dirname, fileName), 'utf8', function(err, data) {
         if (err) {
@@ -91,15 +136,16 @@ function read(dirname, fileName) {
         return output;
     })
 }
+*/
 
 //export all function as amodule
 module.exports = {
     create: create,
     write: write,
-    deleteFile: deleteFile,
-    readFor: readFor,
+    //deleteFile: deleteFile,
+    //readFor: readFor,
     exists: exists,
-    copy: copy,
+    //copy: copy,
     rename: rename,
-    read: read
+    //read: read
 };
